% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{cmap_extract}
\alias{cmap_extract}
\title{Extract Data from CMap GCTX Files for Specified Combinations}
\description{
Functions to extract expression data from CMap GCTX files based on
specified combinations of time points, dosages, and cell lines.
}
\keyword{internal}
