% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_siginfo_beta.R
\name{explore_siginfo_beta}
\alias{explore_siginfo_beta}
\title{Explore siginfo_beta file structure}
\usage{
explore_siginfo_beta(siginfo_file, n_examples = 5)
}
\arguments{
\item{siginfo_file}{Path to siginfo_beta.txt file}

\item{n_examples}{Number of example values to show for each column (default: 5)}
}
\value{
Invisibly returns the column information
}
\description{
This function provides a quick overview of the siginfo_beta file structure,
showing available columns and sample values.
}
\examples{
\dontrun{
explore_siginfo_beta("path/to/siginfo_beta.txt")
}

}
