% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{get_input_with_default}
\alias{get_input_with_default}
\title{Get input with a default value}
\usage{
get_input_with_default(prompt, default)
}
\arguments{
\item{prompt}{Text prompt to display}

\item{default}{Default value to use if no input is provided}
}
\value{
User input or default value
}
\description{
Get input with a default value
}
