% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{get_rid}
\alias{get_rid}
\title{Get Gene IDs and Gene Names for Landmark Genes}
\usage{
get_rid(geneinfo_df, landmark = TRUE)
}
\arguments{
\item{geneinfo_df}{Data frame containing gene information}
}
\value{
List with rid (gene IDs) and genenames (gene symbols)
}
\description{
Get Gene IDs and Gene Names for Landmark Genes
}
\keyword{internal}
