% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{interactive_cmap_workflow}
\alias{interactive_cmap_workflow}
\title{Interactive CMap analysis workflow}
\usage{
interactive_cmap_workflow()
}
\value{
Invisibly returns NULL
}
\description{
Provides an interactive command-line interface to the CMap analysis workflow.
}
