% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_matching.R
\name{print.cmap_signature_result}
\alias{print.cmap_signature_result}
\title{Print method for cmap_signature_result objects}
\usage{
\method{print}{cmap_signature_result}(x, ...)
}
\arguments{
\item{x}{A cmap_signature_result object}

\item{...}{Additional arguments (not used)}
}
\value{
x invisibly
}
\description{
Print method for cmap_signature_result objects
}
