% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{process_combination}
\alias{process_combination}
\title{Process a Single Combination of Time, Dose, and Cell Line}
\usage{
process_combination(
  combination,
  rid,
  genenames,
  sig_info,
  gctx_file = "level5_beta_trt_cp_n720216x12328.gctx",
  output_dir = "."
)
}
\arguments{
\item{combination}{Data frame row containing time, dose, and cell line information}

\item{rid}{Vector of gene IDs to extract}

\item{genenames}{Vector of gene names corresponding to the gene IDs}

\item{sig_info}{Data frame containing signature information}

\item{gctx_file}{Path to the GCTX file}

\item{output_dir}{Directory to save output files}
}
\value{
Path to the output file (invisibly)
}
\description{
Process a Single Combination of Time, Dose, and Cell Line
}
\keyword{internal}
