% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{select_items_interactive}
\alias{select_items_interactive}
\title{Select items from a list interactively}
\usage{
select_items_interactive(all_items, item_type)
}
\arguments{
\item{all_items}{Character vector of all available items}

\item{item_type}{String describing the type of items}
}
\value{
Character vector of selected items
}
\description{
Select items from a list interactively
}
