% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{select_methods_interactive}
\alias{select_methods_interactive}
\title{Interactive selection of signature matching methods}
\usage{
select_methods_interactive()
}
\value{
Vector of selected method names
}
\description{
Interactive selection of signature matching methods
}
