% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_matching.R
\name{summary.cmap_signature_result}
\alias{summary.cmap_signature_result}
\title{Summary method for cmap_signature_result objects}
\usage{
\method{summary}{cmap_signature_result}(object, top_n = 10, ...)
}
\arguments{
\item{object}{A cmap_signature_result object}

\item{top_n}{Number of top compounds to show (default: 10)}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame with the top compounds
}
\description{
Summary method for cmap_signature_result objects
}
