% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_combinations.R
\name{write_config_file}
\alias{write_config_file}
\title{Write CMap options to configuration file}
\usage{
write_config_file(options_list, file_path)
}
\arguments{
\item{options_list}{List with options (times, doses, cells)}

\item{file_path}{Path to output file}
}
\value{
Path to the written file (invisibly)
}
\description{
Write CMap options to configuration file
}
\keyword{internal}
